requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
bson==0.5.8
cffi==1.14.0
cryptography==2.8
dataclasses==0.6
Django==3.0.4
django-rest-knox==4.1.0
djangorestframework==3.11.0
djongo==1.3.1
dnspython==1.16.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
